Create Java_Program_for_Day_of_the_Week_Calculation_Using_the_Calendar_Class.java
The Calendar class in Java is an abstract class that provides methods for converting between a specific instant in time and a set of calendar fields such as year, month, day of month, hour, minute, and second. It also allows for manipulation of these calendar fields. The Calendar class is used for date and time manipulation in Java applications.

As it is an Abstract class, so we cannot use a constructor to create an instance. Instead, we will have to use the static method Calendar.getInstance() to instantiate and implement a sub-class.

Calendar.getInstance(): return a Calendar instance based on the current time in the default time zone with the default locale.